Revert "ExtensionProvidedDependency"
This reverts commit c830e0c58179d0b2c9ff1d6d81be7c03cb6bdc84.
